refactor: extract IBinaryDiffer to Core — fix PatchMiddleware circular dependency
- Created IBinaryDiffer in Core (GeneralUpdate.Core.Differential namespace)
- Differential's IBinaryDiffer now inherits from Core interface (backward-compat)
- Added .BinaryDiffer<T>() extension point to AbstractBootstrap
- PatchMiddleware now accepts IBinaryDiffer via constructor injection
  and performs real differential patching when a differ is provided
- Without injection, PatchMiddleware gracefully skips with a log message

Closes #346
