+++ src/c#/GeneralUpdate.Core/Configuration/AbstractBootstrap.cs
@@ -83,6 +83,9 @@ namespace GeneralUpdate.Core.Configuration
         public TBootstrap SslPolicy<T>() where T : Security.ISslValidationPolicy, new()
         { _extensions[typeof(Security.ISslValidationPolicy)] = typeof(T); return (TBootstrap)this; }
 
+        public TBootstrap BinaryDiffer<T>() where T : Differential.IBinaryDiffer, new()
+        { _extensions[typeof(Differential.IBinaryDiffer)] = typeof(T); return (TBootstrap)this; }
+
         public TBootstrap DownloadPolicy<T>() where T : Download.Abstractions.IDownloadPolicy, new()
         { _extensions[typeof(Download.Abstractions.IDownloadPolicy)] = typeof(T); return (TBootstrap)this; }
